Improve safetey of brn resources
This change adds regex validators to semver strings in the BRN. It has
also refactored the `node` module to logically separate behaviors
associated with node models.
