+++ models/Cargo.toml
@@ -12,8 +12,10 @@ chrono = "0.4"
 k8s-openapi = { version = "0.13.1", default-features = false, features = ["v1_20"] }
 kube = { version = "0.62.0", default-features = true, features = [ "derive"] }
 
+lazy_static = "1.4"
 maplit = "1.0"
 mockall = { version = "0.10", optional = true }
+regex = "1.5"
 schemars = "0.8"
 semver = "1.0"
 serde = { version = "1", features = [ "derive" ] }
@@ -21,3 +23,4 @@ serde_json = "1"
 snafu = "0.6"
 tokio = { version = "1", features = ["macros", "rt-multi-thread", "time"] }
 tracing = "0.1"
+validator = { version = "0.14", features = ["derive"] }
